requirements.txt
asgiref==3.3.1
boto3==1.17.43
botocore==1.20.43
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-headers==3.7.0
django-treenode==0.15.0
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
geographiclib==1.50
geopy==2.1.0
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
twilio==6.51.1
urllib3==1.26.4
django-easy-softdelete==0.1.0
pyfcm==1.5.1
stripe==2.58.0
django-storages==1.12.1